refactor: reviewDto로 이름 변경, service에서 user 검증 절차 추가

+++ src/main/java/team03/mopl/domain/review/service/ReviewServiceImpl.java
@@ -6,12 +6,14 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import team03.mopl.common.exception.content.ContentNotFoundException;
+import team03.mopl.common.exception.review.ReviewDeleteDeniedException;
 import team03.mopl.common.exception.review.ReviewNotFoundException;
+import team03.mopl.common.exception.review.ReviewUpdateDeniedException;
 import team03.mopl.common.exception.user.UserNotFoundException;
 import team03.mopl.domain.content.Content;
 import team03.mopl.domain.content.repository.ContentRepository;
 import team03.mopl.domain.review.dto.ReviewCreateRequest;
-import team03.mopl.domain.review.dto.ReviewResponse;
+import team03.mopl.domain.review.dto.ReviewDto;
 import team03.mopl.domain.review.dto.ReviewUpdateRequest;
 import team03.mopl.domain.review.entity.Review;
 import team03.mopl.domain.review.repository.ReviewRepository;
@@ -28,15 +30,15 @@ public class ReviewServiceImpl implements ReviewService {
   private final ContentRepository contentRepository;
 
   @Override
-  public ReviewResponse create(ReviewCreateRequest request) {
+  public ReviewDto create(ReviewCreateRequest request) {
 
     if (!userRepository.existsById(request.userId())) {
-      log.debug("존재하지 않는 유저입니다. 유저 ID: ", request.userId());
+      log.warn("존재하지 않는 유저입니다. 유저 ID: ", request.userId());
       throw new UserNotFoundException();
     }
 
     if (!contentRepository.existsById(request.contentId())) {
-      log.debug("존재하지 않는 콘텐츠입니다. 콘텐츠 ID: ", request.contentId());
+      log.warn("존재하지 않는 콘텐츠입니다. 콘텐츠 ID: ", request.contentId());
       throw new ContentNotFoundException();
     }
 
@@ -49,55 +51,58 @@ public class ReviewServiceImpl implements ReviewService {
         .comment(request.comment())
         .rating(request.rating())
         .build();
-    return ReviewResponse.from(reviewRepository.save(review));
+    return ReviewDto.from(reviewRepository.save(review));
   }
 
   @Override
-  public ReviewResponse update(UUID reviewId, ReviewUpdateRequest request) {
+  public ReviewDto update(UUID reviewId, ReviewUpdateRequest request, UUID userId) {
     Review review = reviewRepository.findById(reviewId).orElseThrow(ReviewNotFoundException::new);
+    if (!review.getUser().getId().equals(userId)) {
+      log.warn("리뷰 작성자만 수정할 수 있습니다. 리뷰 작성자 ID: ", userId);
+      throw new ReviewUpdateDeniedException();
+    }
 
     review.update(request.title(), request.comment(), request.rating());
-    return ReviewResponse.from(reviewRepository.save(review));
+    return ReviewDto.from(reviewRepository.save(review));
   }
 
   @Override
-  public ReviewResponse get(UUID reviewId) {
+  public ReviewDto get(UUID reviewId) {
     Review review = reviewRepository.findById(reviewId).orElseThrow(ReviewNotFoundException::new);
 
-    return ReviewResponse.from(review);
+    return ReviewDto.from(review);
   }
 
   @Override
-  public List<ReviewResponse> getAllByUser(UUID userId) {
+  public List<ReviewDto> getAllByUser(UUID userId) {
     if (!userRepository.existsById(userId)) {
-      log.debug("존재하지 않는 유저입니다. 유저 ID: ", userId);
+      log.warn("존재하지 않는 유저입니다. 유저 ID: ", userId);
       throw new UserNotFoundException();
     }
 
     List<Review> reviewList = reviewRepository.findAllByUserId(userId);
-    return reviewList.stream().map(ReviewResponse::from).toList();
+    return reviewList.stream().map(ReviewDto::from).toList();
   }
 
   @Override
-  public List<ReviewResponse> getAllByContent(UUID contentId) {
+  public List<ReviewDto> getAllByContent(UUID contentId) {
     if (!contentRepository.existsById(contentId)) {
-      log.debug("존재하지 않는 콘텐츠입니다. 콘텐츠 ID: ", contentId);
+      log.warn("존재하지 않는 콘텐츠입니다. 콘텐츠 ID: ", contentId);
       throw new ContentNotFoundException();
     }
 
     List<Review> reviewList = reviewRepository.findAllByContentId(contentId);
-    return reviewList.stream().map(ReviewResponse::from).toList();
+    return reviewList.stream().map(ReviewDto::from).toList();
   }
 
   @Override
-  public void delete(UUID reviewId) {
+  public void delete(UUID reviewId, UUID userId) {
     Review review = reviewRepository.findById(reviewId).orElseThrow(ReviewNotFoundException::new);
-    reviewRepository.delete(review);
-  }
+    if (!review.getUser().getId().equals(userId)) {
+      log.warn("리뷰 작성자만 삭제할 수 있습니다. 리뷰 작성자 ID: ", userId);
+      throw new ReviewDeleteDeniedException();
+    }
 
-  @Override
-  public void deleteAllByUser(UUID userId) {
-    User user = userRepository.findById(userId).orElseThrow(UserNotFoundException::new);
-    reviewRepository.deleteAllByUserId(userId);
+    reviewRepository.delete(review);
   }
 }
